# src/models/commandregistry.ts
// Generated by SpecForge — do not edit manually
// Model: CommandRegistry

export interface CommandRegistry {
  id: string;
  commands: Record<string, unknown>;
  configPath: string;
  loadedAt: Date;
}

export interface CreateCommandRegistryInput {
  commands: Record<string, unknown>;
  configPath?: string;
}

export interface UpdateCommandRegistryInput {
  commands?: Record<string, unknown>;
  configPath?: string;
}