# src/models/commandvariable.ts
// Generated by SpecForge — do not edit manually
// Model: CommandVariable

export interface CommandVariable {
  id: string;
  name: string;
  required: boolean;
  default: string;
  description: string;
  commandId: string;
}

export interface CreateCommandVariableInput {
  name: string;
  required?: boolean;
  default: string;
  description: string;
  commandId: string;
}

export interface UpdateCommandVariableInput {
  name?: string;
  required?: boolean;
  default?: string;
  description?: string;
  commandId?: string;
}