# src/models/customcommand.ts
// Generated by SpecForge — do not edit manually
// Model: CustomCommand

export interface CustomCommand {
  id: string;
  name: string;
  description: string;
  run: string;
  alias: string;
  variables: Record<string, unknown>;
  createdAt: Date;
  updatedAt: Date;
}

export interface CreateCustomCommandInput {
  name: string;
  description: string;
  run: string;
  alias: string;
  variables: Record<string, unknown>;
}

export interface UpdateCustomCommandInput {
  name?: string;
  description?: string;
  run?: string;
  alias?: string;
  variables?: Record<string, unknown>;
}