# src/models/post.ts
// Generated by SpecForge — do not edit manually
// Model: Post

export interface Post {
  id: string;
  title: string;
  body: string;
  published: boolean;
  authorId: string;
}

export interface CreatePostInput {
  title: string;
  body: string;
  published?: boolean;
  authorId: string;
}

export interface UpdatePostInput {
  title?: string;
  body?: string;
  published?: boolean;
  authorId?: string;
}