# src/models/user.ts
// Generated by SpecForge — do not edit manually
// Model: User

export type roleType = "admin" | "user" | "moderator";

export interface User {
  id: string;
  email: string;
  name: string;
  role: "admin" | "user" | "moderator";
  createdAt: Date;
}

export interface CreateUserInput {
  email: string;
  name: string;
  role?: "admin" | "user" | "moderator";
}

export interface UpdateUserInput {
  email?: string;
  name?: string;
  role?: "admin" | "user" | "moderator";
}