# src/routes/customcommand.ts
// Generated by SpecForge — do not edit manually
// Routes for CustomCommand (/commands)

import type { FastifyInstance } from "fastify";

export default async function customCommandRoutes(app: FastifyInstance) {
  // List all customcommands
  app.get("/commands", async (request, reply) => {
    // TODO: Implement list customcommands
    return { data: [] };
  });

  // Get a single customcommand by ID
  app.get("/commands/:id", async (request, reply) => {
    const { id } = request.params as { id: string };
    // TODO: Implement get customcommand by id
    return { data: null };
  });

  // Create a new customcommand
  app.post("/commands", async (request, reply) => {
    const body = request.body;
    // TODO: Implement create customcommand
    reply.status(201);
    return { data: body };
  });

  // Update a customcommand
  app.put("/commands/:id", async (request, reply) => {
    const { id } = request.params as { id: string };
    const body = request.body;
    // TODO: Implement update customcommand
    return { data: { id, ...body as object } };
  });

  // Delete a customcommand
  app.delete("/commands/:id", async (request, reply) => {
    const { id } = request.params as { id: string };
    // TODO: Implement delete customcommand
    reply.status(204);
    return;
  });

}