# src/routes/post.ts
// Generated by SpecForge — do not edit manually
// Routes for Post (/posts)

import type { FastifyInstance } from "fastify";

export default async function postRoutes(app: FastifyInstance) {
  // List all posts
  app.get("/posts", async (request, reply) => {
    // TODO: Implement list posts
    return { data: [] };
  });

  // Get a single post by ID
  app.get("/posts/:id", async (request, reply) => {
    const { id } = request.params as { id: string };
    // TODO: Implement get post by id
    return { data: null };
  });

  // Create a new post
  app.post("/posts", async (request, reply) => {
    const body = request.body;
    // TODO: Implement create post
    reply.status(201);
    return { data: body };
  });

  // Update a post
  app.put("/posts/:id", async (request, reply) => {
    const { id } = request.params as { id: string };
    const body = request.body;
    // TODO: Implement update post
    return { data: { id, ...body as object } };
  });

  // Delete a post
  app.delete("/posts/:id", async (request, reply) => {
    const { id } = request.params as { id: string };
    // TODO: Implement delete post
    reply.status(204);
    return;
  });

}