# src/routes/user.ts
// Generated by SpecForge — do not edit manually
// Routes for User (/users)

import type { FastifyInstance } from "fastify";

export default async function userRoutes(app: FastifyInstance) {
  // List all users
  app.get("/users", async (request, reply) => {
    // TODO: Implement list users
    return { data: [] };
  });

  // Get a single user by ID
  app.get("/users/:id", async (request, reply) => {
    const { id } = request.params as { id: string };
    // TODO: Implement get user by id
    return { data: null };
  });

  // Create a new user
  app.post("/users", async (request, reply) => {
    const body = request.body;
    // TODO: Implement create user
    reply.status(201);
    return { data: body };
  });

  // Update a user
  app.put("/users/:id", async (request, reply) => {
    const { id } = request.params as { id: string };
    const body = request.body;
    // TODO: Implement update user
    return { data: { id, ...body as object } };
  });

  // Delete a user
  app.delete("/users/:id", async (request, reply) => {
    const { id } = request.params as { id: string };
    // TODO: Implement delete user
    reply.status(204);
    return;
  });

}